refactor modo module + apply same structure for question and mention
