+++ src/utils/observer.ts
@@ -1,7 +1,6 @@
-import { addClassMention } from "@/modules/mentions/mention.module";
-import { getActualUserName } from "@/modules/users/user.module";
-import { addClassModerator } from "@/modules/moderator";
-import { addClassQuestion } from "@/modules/questions";
+import { checkForMentions } from "@/modules/mentions/mention.module";
+import { checkForModeratorMessages } from "@/modules/moderators/moderator.module";
+import { checkForQuestions } from "@/modules/question/question.module";
 
 export const observer = new MutationObserver((mutations) => {
   for (const mutation of mutations) {
@@ -16,31 +15,18 @@ export function checkNewMessages() {
   console.log('[WWSNB] Start checking new messages for questions');
   // Get all messages using data-test attribute
   const messages = document.querySelectorAll('[data-test="chatUserMessageText"]') as unknown as HTMLElement[];
-  const actualUserName = getActualUserName();
 
   for (const message of messages) {
     const textContent = message.textContent;
 
-    /**
-     * Check new messages for questions 
-     */        
-    if (textContent?.includes('@question')) addClassQuestion(message);
-    /**
-     * Check new messages for questions and mentions
-     */
-    if (textContent?.includes(`@${actualUserName}`)) addClassMention(message);
-    /**
-     * Check new messages for moderator messages
-     */
-    if (message.hasAttribute('moderator-checked')) {
-      continue;
-    }
-    message.setAttribute('moderator-checked', 'true');
-    
-    // Look for moderator avatar
-    const moderatorAvatar = message.querySelector('[data-test="moderatorAvatar"]');
+    // Check for questions
+    checkForQuestions(message, textContent);
 
-    if (moderatorAvatar) addClassModerator(message);
+    // Check for mentions
+    checkForMentions(message, textContent);
+
+    // Check for moderator messages
+    checkForModeratorMessages(message);
 
 
   }
